+++ changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.24] - 2020-04-09
+
 ### Added
 
 -   ability to disable redirect following when pjaxing pages using the `followRedirects` config variable [#61](https://github.com/Pageworks/djinnjs/issues/61)
@@ -317,7 +319,8 @@ import { hookup, disconnect, message } from "djinnjs/broadcaster";
     -   Base DjinnJS configuration file
     -   Intial DjinnJS compiler scripts
 
-[unreleased]: https://github.com/pageworks/djinnjs/compare/v0.0.23...HEAD
+[unreleased]: https://github.com/pageworks/djinnjs/compare/v0.0.24...HEAD
+[0.0.24]: https://github.com/pageworks/djinnjs/compare/v0.0.23...v0.0.24
 [0.0.23]: https://github.com/pageworks/djinnjs/compare/v0.0.22...v0.0.23
 [0.0.22]: https://github.com/pageworks/djinnjs/compare/v0.0.21...v0.0.22
 [0.0.21]: https://github.com/pageworks/djinnjs/compare/v0.0.20...v0.0.21